Audiobook-shelf-master/Naver AuidoClip Crawling/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
attrs==19.3.0
awscli==1.18.5
Babel==2.8.0
beautifulsoup4==4.8.2
black==19.10b0
boto3==1.12.5
botocore==1.15.5
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
check-manifest==0.42
click==7.1.2
colorama==0.4.3
coverage==5.1
docutils==0.15.2
idna==2.9
imagesize==1.2.0
jarn.viewdoc==2.3
Jinja2==2.11.2
jmespath==0.9.4
MarkupSafe==1.1.1
more-itertools==8.3.0
olefile==0.46
packaging==20.4
pathspec==0.8.0
pep517==0.8.2
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pyroma==2.6
pytest==5.4.3
pytest-cov==2.9.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.2
regex==2020.6.8
requests==2.23.0
rsa==3.4.2
s3transfer==0.3.3
schedule==0.6.0
selenium==3.141.0
six==1.14.0
snowballstemmer==2.0.0
soupsieve==1.9.5
Sphinx==3.1.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.2.4
